move state extraction helpers to stateextractor.go
